datatransfer/pyproject.toml
[tool.poetry]
name = "datatransfer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
asgiref = "3.4.1"
cachetools = "4.2.2"
certifi = "2021.5.30"
cffi = "1.14.6"
charset-normalizer = "2.0.4"
Django = "3.2.5"
django-debug-toolbar = "3.2.1"
django-filter = "2.4.0"
djangorestframework = "3.12.4"
google-api-core = "1.31.1"
google-auth = "1.34.0"
google-cloud-core = "1.7.2"
google-cloud-storage = "1.41.1"
google-crc32c = "1.1.2"
google-resumable-media = "1.3.3"
googleapis-common-protos = "1.53.0"
idna = "3.2"
packaging = "21.0"
protobuf = "3.17.3"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycparser = "2.20"
pyparsing = "2.4.7"
pytz = "2021.1"
requests = "2.26.0"
rsa = "4.7.2"
six = "1.16.0"
sqlparse = "0.4.1"
urllib3 = "1.26.6"
mysqlclient = "^2.0.3"
gunicorn = "^20.1.0"
pydantic = "^1.8.2"
python-dotenv = "^0.19.0"
drf-spectacular = "^0.20.2"
opencensus = "^0.8.0"
opencensus-ext-django = "^0.7.6"
opencensus-ext-azure = "^1.1.0"
opencensus-ext-requests = "^0.7.6"
opencensus-ext-google-cloud-clientlibs = "^0.1.3"
structlog = "^21.2.0"
uvicorn = "^0.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"